requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=3.14
django-filter>=23.2
drf-yasg>=1.21
python-dotenv>=1.0.0
